Require key and projectId in createClient()
This change adds a check to ensure createClient() gets a key and
projectId passed into the createClient() call.

In previous versions of Airbrake the first argument was the key,
meaning that blindly upgrading the Airbrake version would result in
Airbrake throwing an error when reporting an error.

This change adds requirements that both the key and projectId be
supplied, "failing fast" if either of these are omitted.  This
gives the developer feedback on this quickly, rather than have
the Airbrake library throw errors when reporting errors to the
service.

Fixes #124