truncator: use 'JSON.stringify' instead of 'json-stringify-safe'
Possibly fixes #131 (JSON parsing problem with error messages involving
a pound sterling symbol (£))

We used to use this library when we didn't have payload truncation
support. Now, when payload truncation is in place, we truncate notices
before sending. The truncator resolves circular references, so we don't
need to depend on 'json-stringify-safe' anymore.